refactor: move dynamically loaded component into it's own file
